`Key` option and TESSEL_AUTH_KEY
As mentioned by @johnnyman727 at
https://github.com/tessel/t2-cli/pull/290/files#r39475328 and liked by
@wprater at https://github.com/tessel/t2-cli/pull/328/files#r39893445 I
am going to make a suggestion how to do by this pull request.

### Notes
It was necessary to change a test about `update` command (assume
@rwaldron wants to check this).

### Usage
Each command what likes to have a default value for the `option('key')`
would be able to use `TESSEL_AUTH_KEY`.