Rename Tessel.(RUN|PUSH)_PATH => REMOTE_(RUN|PUSH)_PATH; Rename Tessel.TESSEL_AUTH_(PATH|KEY) => Tessel.LOCAL_AUTH_(PATH|KEY)
Previous there were 4 file/path-centric "const-ish" properties defined on the Tessel class, two of which were specific to some local operations and two of which were specific to some remote operations. None of the 4 values clearly indicated which of those operations they were relevant to. With this renaming, it becomes very clear:

- Tessel.RUN_PATH => Tessel.REMOTE_RUN_PATH
- Tessel.PUSH_PATH => Tessel.REMOTE_PUSH_PATH
- Tessel.TESSEL_AUTH_PATH => Tessel.LOCAL_AUTH_PATH
- Tessel.TESSEL_AUTH_KEY => Tessel.LOCAL_AUTH_KEY