Deploy: further refine handling of "unparseable" sources. Fixes gh-587
The algorithm defined in https://github.com/tessel/t2-cli/commit/e6639b4fd9508269dc8f4a38b0a6bc437accbd4f was an improvement, but it's clear that we'll play "whack-a-mole" forever if we attempt to hand-hold too much. After reading #587, the obvious step was to simply NOT compress code that could not be parsed by uglify or acorn. If the code truly contains a syntax error, let the runtime itself handle that.

Taking this approach, which relaxes the potential for failures also addresses the outstanding issue with ES6 syntax being unparseable by acorn and/or uglify. Now, if either fail to parse because of an arrow function (or anything), the deployment is no longer terminated.