Wifi: eliminate use of timeout in wifi connection verification
Eliminated the time out and replace it with a recursively called "polling" mechanism, I could increase the likelihood of successfully verifying the wifi connection.

Arbitrarily selected 10 as the limit of tries that will be made; this could actually be increased to ~33, which is the approximate number of tries that could potentially be made in 12 seconds (the delay period (2) + timeout period (10)). I determined this by counting the number of complete tries made against a known invalid ssid, within a 12 second period. I ran the measurement code 10 times and the rounded average count was 33 tries.